+++ libs/ledger-live-common/src/families/tron/deviceTransactionConfig.ts
@@ -15,7 +15,7 @@ export type ExtraDeviceTransactionField =
     };
 
 function getDeviceTransactionConfig({
-  transaction: { votes, resource, mode },
+  transaction: { votes, resource, mode, recipient },
   account,
   parentAccount,
   status: { amount },
@@ -59,11 +59,27 @@ function getDeviceTransactionConfig({
     });
   }
 
-  if (mode === "unfreeze") {
+  if (mode === "unfreeze" || mode === "legacyUnfreeze") {
     fields.push({
       type: "address",
-      label: "Delegated To",
+      label: "Unfreeze To",
       address: mainAccount.freshAddress,
+    });
+  }
+
+  if (mode === "withdrawExpireUnfreeze") {
+    fields.push({
+      type: "address",
+      label: "Withdraw unfrozen to",
+      address: mainAccount.freshAddress,
+    });
+  }
+
+  if (mode === "unDelegateResource") {
+    fields.push({
+      type: "address",
+      label: "Undelegate from",
+      address: recipient,
     });
   }