+++ libs/ledgerjs/packages/hw-app-vet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-vet
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`d49f444`](https://github.com/LedgerHQ/ledger-live/commit/d49f44417fd175affe71da589c0ca380e88fbb35)]:
+  - @ledgerhq/cryptoassets@11.3.1
+
+## 0.1.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`d49f444`](https://github.com/LedgerHQ/ledger-live/commit/d49f44417fd175affe71da589c0ca380e88fbb35)]:
+  - @ledgerhq/cryptoassets@11.3.1-next.0
+
 ## 0.1.0
 
 ### Minor Changes